+++ pages/Newqa.js
@@ -5,7 +5,7 @@ import { getFirestore, collection, addDoc, setDoc,doc,Timestamp,getDoc} from "fi
 import { updateDoc, serverTimestamp } from "firebase/firestore";
 import {firebaseConfig} from '../settings/firebaseConfig';
 import { query, orderBy, limit } from "firebase/firestore";
-import {Container, AppBar, Box, Toolbar, IconButton, Typography, Button, InputBase } from '@mui/material'
+import {Container, AppBar, Box, Toolbar, IconButton, Typography, Button, InputBase, Card, CardActions } from '@mui/material'
 import { styled, alpha } from '@mui/material/styles'
 import Navbar from "../components/navbar/Navbar";
 import styles from '../styles/Home.module.css';
@@ -25,46 +25,74 @@ import InputLabel from '@mui/material/InputLabel';
 import MenuItem from '@mui/material/MenuItem';
 import FormControl from '@mui/material/FormControl';
 import Select, { SelectChangeEvent } from '@mui/material/Select';
+import FormHelperText from '@mui/material/FormHelperText';
 
 
 const MENU_LIST = [
-  { text: "登入", href: "/login" },
-  //{ text: "註冊", href: "/register"},
-];
-
-    // const [navActive, setNavActive] = useState(null);
-    // const [activeIdx, setActiveIdx] = useState(-1);
+    { text: "登入", href: "/login" },
+    //{ text: "註冊", href: "/register"},
+  ];
   
-    const firebaseApp = getApps().length === 0 ? initializeApp(firebaseConfig) : getApp();
-    const db = getFirestore();
-
-    const auth = getAuth();
-    
-function Newqa () {
+  // const [navActive, setNavActive] = useState(null);
+  // const [activeIdx, setActiveIdx] = useState(-1);
+  
+  const firebaseApp = getApps().length === 0 ? initializeApp(firebaseConfig) : getApp();
+  const db = getFirestore();
+  
+  const auth = getAuth();
+  
+  function Newqa() {
+  
     const router = useRouter();
+    const { articleId } = router.query;
     const [title, setTitle] = React.useState('');
     const [content, setContent] = React.useState('');
     const [tags, setTags] = React.useState([]);
     const [tagName, setTagName] = React.useState("");
     const [link, setLink] = React.useState('');
-    const [age, setAge] = React.useState('');
     const [user, setUser] = useState();
-
-
-
-
-
+  
+  
     useEffect(() => {
-        const unsub = onAuthStateChanged(auth, (user)=>{
-          setUser(user);
-          console.log(user);
-        });
-    
-        return () => {
-          unsub();
+  
+      async function readData() {
+        if (articleId) {
+          const ref = doc(db, "text", articleId);
+          const docSnapshot = await getDoc(ref);
+          if (docSnapshot.exists()) {
+            console.log("doc", docSnapshot.data())
+            setTitle(docSnapshot.data().title)
+            setContent(docSnapshot.data().content)
+            setLink(docSnapshot.data().link)
+            setTagName(docSnapshot.data().tag)
+          }
         }
-      }, []);
-
+  
+      }
+      readData();
+    }
+      , [articleId])
+  
+  
+  
+    useEffect(() => {
+      //   if (articleId){
+      //   const ref = doc(db, "text", articleId);
+      //   const docSnapshot = await getDoc(ref);
+      //   if (docSnapshot.exists()) {
+      //   setTitle(docSnapshot.data().title)
+      //   }
+      // }
+      const unsub = onAuthStateChanged(auth, (user) => {
+        setUser(user);
+        console.log(user);
+      });
+  
+      return () => {
+        unsub();
+      }
+    }, []);
+  
     // React.useState(() =>{
     //     firebase
     //     .firestore
@@ -76,161 +104,170 @@ function Newqa () {
     //     });
     //     setTopics(data);
     // });
-
+  
     // const handleChange = event => {
     //     console.log(event.target.value);
     //     setSelected(event.target.value);
     //   };
-      
-
+  
+  
     const options = tags.map(tag => {
-        return {
-            text:tag.name,
-            value:tag.name
-        }
+      return {
+        text: tag.name,
+        value: tag.name
+      }
     })
-
-    
-
-    async function onSubmit(){
-
-        // console.log(tagName);
-        // alert(user.uid)
-        // alert(user.email)        
-        // await addDoc(collection(db, "text",
-        // props.article.docId,"comment"))
-        await addDoc(collection(db, "question"), {
+  
+  
+  
+  
+    const update = async function () {
+      if (title == "" || content == "" || tagName == "" || link == "") {
+        return (false);
+      }
+  
+      const db = getFirestore();
+      try {
+        if (!articleId) {
+          const docRef = await addDoc(collection(db, "text"), {
             title,
             content,
             userid: user.uid,
             email: user.email,
-            tag:tagName,
-            user:user.displayName,
-            heart:[],
-            bookmark:[],
-            //還有這些無法加入生成欄位，看來是需要給一個值嗎？
-            count:1,
-            // link,
-            //user欄位要帶入登入的資料8
-            //user, 
+            tag: tagName,
+            user: user.displayName,
+            heart: [],
+            bookmark: [],
+            count: 1,
             link,
+            outdate: "solved",
+            outdateCount: [],
             timestamp: serverTimestamp()
-            // createAt:Timestamp.now(),
-            // author:{
-            //     displayName: auth.currentUser.displayName || "",
-            //     photoURL: auth.currentUser.photoURL || "",
-            //     uid: auth.currentUser.uid,
-            //     email: auth.currentUser.email
-            // },
           });
-        
-     
-            router.push('/');
-        
+          console.log(docRef.id);
+        }
+        else {
+          await updateDoc(doc(db, "text", articleId), {
+            title,
+            content,
+            tag: tagName,
+            link,
+          });
+        }
+      }
+      catch (e) {
+        console.log(e);
+      }
+      router.push('/');
     }
 
-    // function MultilineTextFields() {
-    //     return (
-    //       <Box
-    //         component="form"
-    //         sx={{
-    //           '& .MuiTextField-root': { m: 1, width: '80ch' },
-    //         }}
-    //         noValidate
-    //         autoComplete="off"
-    //       >
-    //         {/* <div>
-    //           <TextField
-    //             id="outlined-multiline-flexible"
-    //             label="Multiline"
-    //             multiline
-    //             maxRows={4}
-    //           />
-    //         </div> */}
-    //         </Box>
-    //         );
-    //         }
-
     return (
-    <div className={styles.post_container}>
-        <div>
-        <Navbar/>
-
+    //<div className={styles.post_container}>
+    <div>
         
-        <h3 className>發布問題</h3><br/>
-      
-            <div>
-                {user&&user.displayName}
-            <FormControl fullWidth>
-            <TextField
-                id="outlined-textarea"
-                label="請輸入提問標題"
-                placeholder="想問的是..."
-                multiline
-                onChange={(e) => setTitle(e.target.value)} 
-            /></FormControl>
-            </div><br/>
+        
+    <div>
+        <Navbar/>
+        <Toolbar/>
+        <Container>
+            <Box display="flex" flexDirection="column" flexWrap="wrap" justifyContent="center" >
+                <Card
+                display="flex"
+                flexDirection="column"
+                sx={{
+                    // bgcolor: "#000000",
+                    borderRadius: 2,
+                    p:2,
+                    pr: 4,
+                    pl: 4,
+                    m: 4,
+                }}
+                >
 
-           
-            <div>
-            <FormControl fullWidth>
-            <TextField
-                InputProps={{ sx: { height: 250 } }}
-                id="outlined-textarea"
-                label="請輸入問題內容"
-                placeholder="我想問..."
-                multiline
-                onChange={(e) => setContent(e.target.value)} 
-            /></FormControl>
-            </div><br/>
+                <Typography variant="h6">發布筆記</Typography>
+                <Typography variant="h6">{user && user.displayName}</Typography>
 
-           
-            {/* <div>
-            <FormControl fullWidth>
-            <TextField
-                
-                id="outlined-textarea"
-                label="您想分享的連結"
-                placeholder="https..."
-                multiline
-                onChange={(e) => setLink(e.target.value)} 
-                
-            />
-            </FormControl>
-            </div ><br/> */}
-            <br/>
-            <FormControl sx={{width:250}}>
-            <InputLabel id="demo-simple-select-label">請輸入問題標籤</InputLabel>
-            <Select
-                labelId="demo-simple-select-label"
-                id="demo-simple-select"
-                // value={topicName}
-                // label="topic"
-                onChange={(e) => {setTagName(e.target.value); 
-                console.log("t:")}} 
-            >
-                <MenuItem value="課堂筆記">課堂筆記</MenuItem>
-                <MenuItem value="修課心得">修課心得</MenuItem>
-                <MenuItem value="專題相關">專題相關</MenuItem>
-                <MenuItem value="業界資源">業界資源</MenuItem>
-                <MenuItem value="其他">其他</MenuItem>
-            </Select><br/>
-{/* <InputLabel id="label">Age</InputLabel>
-<Select labelId="label" id="select" value="20">
-  <MenuItem value="20">Ten</MenuItem>
-  <MenuItem value="10">Twenty</MenuItem>
-</Select> */}
+                <FormControl fullWidth>
+                    <TextField
+                     error={title === ""}
+                    //  helperText={title === "" ? "請輸入標題" : ""}
+                     required
+                     id="outlined-textarea"
+                     label="請輸入問題標題"
+                     placeholder={title ? "" : "今天的主題是..."}
+                     multiline
+                     value={title}
+                     margin="normal"
+                     onChange={(e) => setTitle(e.target.value)}
+                    />
+                </FormControl>
+
+                <FormControl fullWidth>
+                <TextField
+                  error={content === ""}
+                //   helperText={content === "" ? "請輸入內容" : ""}
+                  required
+                  InputProps={{ sx: { height: 250 } }}
+                  id="outlined-textarea"
+                  label="請輸入內容"
+                  placeholder={content ? "" : "我想問..."}
+                  margin="normal"
+                  multiline
+                  value={content}
+                  onChange={(e) => setContent(e.target.value)}
+                />
+                </FormControl>
+
+                {/* <FormControl fullWidth>
+                    <TextField
+                    error={link === ""}
+                    helperText={link === "" ? "請輸入連結" : ""}
+                    id="outlined-textarea"
+                    label="您想分享的連結"
+                    placeholder={link ? "" : "https..."}
+                    multiline
+                    value={link}
+                    margin="normal"
+                    onChange={(e) => setLink(e.target.value)}
+
+                    />
+                </FormControl> */}
+
+                {/* <FormControl sx={{ width: 250 }}
+                error={tagName === ""}
+                margin="normal"
+              >
+                <InputLabel id="demo-simple-select-label" required>請選擇問題標籤</InputLabel>
+                <Select
+                  labelId="demo-simple-select-label"
+                  id="demo-simple-select"
+                  value={tagName}
+                  // label="topic"
+                  onChange={(e) => {
+                    setTagName(e.target.value);
+                    console.log("t:")
+                  }}
+                >
+                  <MenuItem value="課堂筆記">課堂筆記</MenuItem>
+                  <MenuItem value="修課心得">修課心得</MenuItem>
+                  <MenuItem value="專題相關">專題相關</MenuItem>
+                  <MenuItem value="業界資源">業界資源</MenuItem>
+                  <MenuItem value="其他">其他</MenuItem>
+                </Select>
+                {tagName === "" && <FormHelperText>請選擇分類標籤</FormHelperText>}
+              </FormControl> */}
+              <CardActions sx={{ display: "flex", justifyContent: "flex-end" }}>
+                <Button variant="contained" onClick={update}>發布</Button><br></br><br></br>
+                <Button variant="contained">取消</Button>
+              </CardActions>
 
 
+            </Card>
+          </Box>
 
-            </FormControl>
 
-          
-
-            <Button variant="contained" sx={{ padding: 1, margin: 5 , left: 680 , top: 40}} onClick={onSubmit}>發布</Button><br></br><br></br>
-            <Button variant="contained" sx={{ padding: 1, margin: 4.5 , left: 1020 , top: -95}}>取消</Button>
-    
-</div>
+        </Container>
+    </div>
            </div>
